refactor: Rename header file extensions to `.hpp`.
Only done it for the ESP32 integration of the `sabre` library for now.
The `sabre` library itself will be done in a different repository and
copied into this one. This one will, eventually, be moved to a
component.

Fixes #45
