+++ components/sabre_esp32/sabre_esp32/factory.cpp
@@ -1,5 +1,5 @@
-#include "factory.h"
-#include "exceptions/exceptions.h"
+#include "factory.hpp"
+#include "exceptions/exceptions.hpp"
 
 namespace sabre::esp32
 {
