+++ components/sabre_esp32/sabre_esp32/factory.hpp
@@ -1,13 +1,13 @@
 #ifndef SABRE_ESP32_FACTORY_H
 #define SABRE_ESP32_FACTORY_H
 
-#include "clients/mqtt.h"
-#include "gpio/input_gpio.h"
-#include "gpio/output_gpio.h"
-#include "uart/uart.h"
-#include "utility/timed_waiter.h"
-#include "wifi/wifi_soft_ap.h"
-#include "wifi/wifi_station.h"
+#include "clients/mqtt.hpp"
+#include "gpio/input_gpio.hpp"
+#include "gpio/output_gpio.hpp"
+#include "uart/uart.hpp"
+#include "utility/timed_waiter.hpp"
+#include "wifi/wifi_soft_ap.hpp"
+#include "wifi/wifi_station.hpp"
 #include <map>
 #include <memory>
 #include <sabre/factory.h>
