+++ components/sabre_esp32/sabre_esp32/gpio/input_gpio.cpp
@@ -1,5 +1,5 @@
-#include "input_gpio.h"
-#include "../exceptions/exceptions.h"
+#include "input_gpio.hpp"
+#include "../exceptions/exceptions.hpp"
 #include <driver/gpio.h>
 
 namespace sabre::esp32
