+++ components/sabre_esp32/sabre_esp32/gpio/output_gpio.cpp
@@ -1,5 +1,5 @@
-#include "output_gpio.h"
-#include "../exceptions/exceptions.h"
+#include "output_gpio.hpp"
+#include "../exceptions/exceptions.hpp"
 #include <driver/gpio.h>
 
 namespace sabre::esp32
