+++ components/sabre_esp32/sabre_esp32/service_runner/service_runner.cpp
@@ -1,4 +1,4 @@
-#include "service_runner.h"
+#include "service_runner.hpp"
 
 namespace sabre::esp32
 {
