+++ components/sabre_esp32/sabre_esp32/uart/uart.cpp
@@ -1,5 +1,5 @@
-#include "uart.h"
-#include "../exceptions/exceptions.h"
+#include "uart.hpp"
+#include "../exceptions/exceptions.hpp"
 
 namespace sabre::esp32
 {
