+++ components/sabre_esp32/sabre_esp32/utility/timed_waiter.cpp
@@ -1,4 +1,4 @@
-#include "timed_waiter.h"
+#include "timed_waiter.hpp"
 
 namespace sabre::esp32
 {
