+++ components/sabre_esp32/sabre_esp32/wifi/wifi_soft_ap.cpp
@@ -1,4 +1,4 @@
-#include "wifi_soft_ap.h"
+#include "wifi_soft_ap.hpp"
 #include <cstring>
 #include <nvs_flash.h>
