+++ components/sabre_esp32/sabre_esp32/wifi/wifi_station.cpp
@@ -1,4 +1,4 @@
-#include "wifi_station.h"
+#include "wifi_station.hpp"
 #include <arpa/inet.h>
 #include <cstring>
 #include <esp_event.h>
